Analysed and processed steering data
The code written so far is based of John Loanes starter code for
this project. When reviewing the steering data it seemed that the
majority of the values present were zero's. Some of these zero
values have now been removed so that the model is not basised
towards not steering.

All other information needed for this project has been read in at
this point it just requires processing before it can be used in
our model.