Analysed and processed throttle data
Analysed the throttle data and as expected there was a lot of
one values contained within the data. These one values were
reduced to prevent a basis towards full throttle.

Removing these values reduced the dataset from 1455 to 852.

Note additional analysis will be required as the dataset grows,
currently working with a small amount of data to setup and
validate the code is working.