Analysed and processed reverse and speed
Analysed and processed the reverse and speed data within the
dataset. Reverse appeared to have a high number of zero values
these were reduced to prevent zero basis. Speed was relatively
normal however, did remove some values from the 30 bin as this
may lead the model to believe that full throttle is favoured.