Add model that runs on both tracks at 30mph
Adding a model that can now run on both tracks successfully at
30 mph - finally. It is worth noting that this is not the first
model that we produced and the hyperparameters experienced many
alterations before this model was discovered.

The correction angle has been updated to 0.2 from 0.15 as this
appears to offer better cornering on the more complicated track.

The Nvidia model while very successful at running the first track
we found was not fully capable at running the second track - at
least with the hyperparameters we provided it. Therefore, we
moved to the VGG16 model, which was original built for extreme
feature extraction from images and scored 92.7% on the ImageNet
database. With this model we seen great improvements.

Moving the the VGG16 model also resulted is us leaving the images
as RGG and resizing the images to 64x64 instead of the original
66X200.

After some manipulation of the learning rate we discovered that
1e-4 was a much more stable learning rate the 1e-3 as this
learning rate appeared to jump across the local minimum
eventually increasing the loss.

We also noticed that when starting the second track the car
was prone to steering around the shadow produced by the sim.
Therefore, to counteract this we added a function called
random_shadow to add random shadows to our images.

Once we had fined tune or model and learning rate, it
appeared that batch size and sample size were two off our
most powerful levers. We decreased the batch size to help
with memory and increased the steps per epoch to increase
the number batches per epoch - this helped lower the cost
greatly.

One interesting point we discovered was that the loss of
the model was not a very good metric for measuring success
in the endeavour to find the perfect model we achieved
models with a loss as low as 0.03 however, these models
appeared to preform worse than models with higher losses
such as 0.06.

I believe the most important change was to our data by adding some
of the more complicated sections of the second track to our data
and recovery data from the second track i.e. recovering from the
right-side of the track to the center or vice versa the model
improved greatly. This allows for the model to correct slight
mistakes when made.

This was a paired programming effort - Aaron and myself worked via
Discord, I programmed and Aaron researched solutions and
different approaches as well as reviewed the code I was writing
and helped add improvements.