Add throttle to working model
The model added via this commit can now drive both tracks with
the throttle being predicted via the model. This allows the
car to brake (throttle 0) and speed up (throttle increases) as it
moves around the track.

In order to achieve this goal some changes had to be made to the
previous model. We increased the steps per epoch and introduced
a new Dense layer to th VGG16 model as well as increasing the
total number of epochs ran to 18.

We also had to present the throttle data to the model for training
this involved combing the steering angle and throttle into an
nx2 dataframe that the model.fit can understand and learn from.

The throttle data was also cleaned in order to balance out braking
and acceleration. This involved removing some of the higher
acceleration values from the data.

Lastly, both the bc.py and drive.py file were cleaned up removing
redundant code. Comments were also added to each of the functions
to help improve the readability of the files.